refactor(function): function 클래스들을 domain/function 패키지로 이동
