+++ src/main/java/subway/domain/function/Functionable.java
@@ -1,8 +1,9 @@
-package subway.domain;
+package subway.domain.function;
 
 import java.util.Arrays;
 import java.util.function.Function;
 
+import subway.domain.ManageController;
 import subway.view.InputView;
 import subway.view.OutputView;
