+++ src/main/java/subway/domain/function/LineFunction.java
@@ -1,6 +1,8 @@
-package subway.domain;
+package subway.domain.function;
 
 import java.util.function.Function;
+
+import subway.domain.ManageController;
 
 public enum LineFunction implements Functionable {
     ADD("1", "노선 등록", ManageController::addLine),
