+++ src/main/java/subway/domain/function/StationFunction.java
@@ -1,13 +1,16 @@
-package subway.domain;
+package subway.domain.function;
 
 import java.util.function.Function;
 
-public enum RangeFunction implements Functionable {
-    ADD("1", "구간 등록", ManageController::addRange),
-    REMOVE("2", "구간 삭제", ManageController::removeRange),
+import subway.domain.ManageController;
+
+public enum StationFunction implements Functionable {
+    ADD("1", "역 등록", ManageController::addStation),
+    REMOVE("2", "역 삭제", ManageController::removeStation),
+    LOAD("3", "역 조회", ManageController::loadStations),
     BACK("B", "돌아가기", Function.identity());
 
-    public static final String TITLE = "구간 관리";
+    public static final String TITLE = "역 관리";
 
     private final String identifier;
 
@@ -15,8 +18,8 @@ public enum RangeFunction implements Functionable {
 
     private final Function<ManageController, ManageController> function;
 
-    RangeFunction(String identifier, String description,
-                  Function<ManageController, ManageController> function) {
+    StationFunction(String identifier, String description,
+                    Function<ManageController, ManageController> function) {
         this.identifier = identifier;
         this.description = description;
         this.function = function;
